start working on image processor
