+++ src/main.rs
@@ -5,6 +5,7 @@ pub mod audio;
 mod cacher;
 pub mod controller;
 pub mod errors;
+pub mod image_processor;
 pub mod library;
 mod queue;
 pub mod scanner;
